migrate d500/safety tests to pytest
