+++ unit-tests/live/d500/safety/pytest-smcu-version.py
@@ -1,0 +1,73 @@
+# License: Apache 2.0. See LICENSE file in root directory.
+# Copyright(c) 2026 RealSense, Inc. All Rights Reserved.
+
+# Not frequently changing, no need to test for each commit
+
+import pytest
+import subprocess
+import re
+import pyrealsense2 as rs
+import pyrsutils as rsutils
+from pytest_check import check
+from rspy import repo
+import logging
+log = logging.getLogger(__name__)
+
+pytestmark = [
+    pytest.mark.device_each("D585S"),
+    pytest.mark.context("nightly"),
+]
+
+
+# Test to read SMCU Version from rsutils FW version, HWMC Commands and rs-fw-update tool,
+# Compare against 0.0.0.0 and check if all the methods provide same result
+def test_smcu_version_consistent_across_methods(test_device):
+    versions_dict = {}
+
+    device, _ = test_device
+
+    ########################  FW version using rs camera info  #######################
+
+    # Get the firmware version using rs_camera_info
+    fw_version = rsutils.version(device.get_info(rs.camera_info.smcu_fw_version))
+
+    # Print the firmware version
+    log.debug("SMCU Version Using rs_camera_info: %s", fw_version)
+    versions_dict["smcu_version_rsutils_fw_version"] = str(fw_version)
+
+    ############################  FW version using HMC  ################################
+
+    # Get the firmware version using GVD cmd
+    dp_device = device.as_debug_protocol()
+    gvd_data = dp_device.send_and_receive_raw_data(dp_device.build_command(0x10))
+    # log.debug ("HMC Command GVD " + str(gvd_data))
+    substring = gvd_data[275:279]
+    hmc_smcu_version = ".".join(str(char) for char in substring)
+    log.debug("SMCU Version Using HMC: %s", hmc_smcu_version)
+    versions_dict["smcu_version_hmc"] = str(hmc_smcu_version)
+
+    ########################## FW version using rs-fw-update  ###########################
+
+    # Get the firmware version using rs-fw-update
+    rs_fw_update_tool_path = repo.find_built_exe('tools/fw-update', 'rs-fw-update')
+    log.debug("rs-fw-update path in the directory : %s", rs_fw_update_tool_path)
+
+    version = subprocess.check_output(rs_fw_update_tool_path).decode("utf-8")
+
+    rs_fw_update_smcu_version = re.search(r'SMCU firmware version: ([\d.]+)', version)
+    if rs_fw_update_smcu_version:
+        smcu_version_rs_fw_update_tool = rs_fw_update_smcu_version.group(1)
+        log.debug("SMCU version using rs-fw-update: %s", smcu_version_rs_fw_update_tool)
+        versions_dict["smcu_version_rs_fw_update"] = str(smcu_version_rs_fw_update_tool)
+
+    log.debug("Actual Dictonary is %s", str(versions_dict.values()))
+
+    # Consider SMCU Versions fetched from one of the methods as reference
+    test_val = versions_dict["smcu_version_rsutils_fw_version"]
+    log.debug("SMCU Version From one of the method = %s", str(test_val))
+    for key, value in versions_dict.items():
+        log.info("%s: %s", key, value)
+        # check the smcu version is not zero
+        check.is_true(value != "0.0.0.0", f"SMCU Version is 0.0.0.0! ({key})")
+        # Check the smcu version is same across multiple methods
+        check.is_true(value == test_val, f"SMCU Version is different than expected! ({key}={value}, expected={test_val})")
